video-14

+++ admin/ajax/features_facilities.php
@@ -38,10 +38,19 @@ if (isset($_POST['get_features'])) {
 if (isset($_POST['rem_feature'])) {
     $frm_data = filteration($_POST);
     $values = [$frm_data['rem_feature']];
-    $q = "DELETE FROM `features` WHERE  `id`=?";
-    $res = delete($q, $values, 'i');
-    echo $res;
-   
+
+    $check_q =select( "SELECT * FROM `room_features` WHERE `features_id`=?",[$frm_data['rem_feature']],'i');
+
+    if(mysqli_num_rows($check_q)==0){
+
+        $q = "DELETE FROM `features` WHERE  `id`=?";
+        $res = delete($q, $values, 'i');
+        echo $res;
+       
+    }else{
+        echo 'room_added';
+    }
+
    
 }
 
@@ -94,6 +103,10 @@ if (isset($_POST['rem_facility'])) {
     $frm_data = filteration($_POST);
     $values = [$frm_data['rem_facility']];
    
+    $check_q =select( "SELECT * FROM `room_facilities` WHERE `facilities_id`=?",[$frm_data['rem_facility']],'i');
+
+    if(mysqli_num_rows($check_q)==0){
+
     $pre_q = "SELECT * FROM `facilities` WHERE `id`=?";
     $res = select($pre_q,$values,'i');
     $img = mysqli_fetch_assoc( $res );
@@ -105,6 +118,10 @@ if (isset($_POST['rem_facility'])) {
     }else{
         echo 0;
     }
+}
+else{
+    echo 'room_added';
+}
    
    
 }
